requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
python-telegram-bot>1.0
apscheduler==3.4.0
